feat: display  icon at the side of code hints. Helps in identifying if hint is coming from emmet
