+++ src/styles/brackets_patterns_override.less
@@ -709,6 +709,18 @@ a:focus {
     }
 }
 
+.emmet-code-hint {
+    position: absolute;
+    font-size: 0.5em;
+    font-weight: bold;
+    right: 4px;
+    bottom: 0px;
+    color: #2ea56c !important;
+    .dark& {
+        color: #146a41 !important;
+    }
+}
+
 #codehint-desc {
 	background: @bc-codehint-desc;
 	position: absolute;
